refact(Widgets): Center genre chips in movie container
