+++ lib/views/widgets/movie_details/movie_details_column.dart
@@ -34,13 +34,10 @@ class MovieDetailsColumn extends HookWidget {
         const SizedBox(height: 15),
 
         //Genres
-        SizedBox(
-          width: 210,
-          child: CustomChipsList(
-            chipHeight: 26,
-            chipGap: 9,
-            chipContents: movie.genreNames.sublist(0, 3),
-          ),
+        CustomChipsList(
+          chipHeight: 26,
+          chipGap: 9,
+          chipContents: movie.genreNames.sublist(0, 3),
         ),
 
         const SizedBox(height: 15),
